Removed clone command
